requirements.txt
altgraph==0.16.1
astroid==2.1.0
auto-py-to-exe==2.5.1
autopep8==1.4.1
beautifulsoup4==4.6.3
bottle==0.12.16
bottle-websocket==0.2.9
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
Eel==0.9.11
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
macholib==1.11
mccabe==0.6.1
pefile==2018.8.8
Pillow==5.4.1
pycodestyle==2.4.0
pycparser==2.19
PyInstaller==3.4
pylint==2.1.1
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyqt5-tools==5.11.3.1.4
python-dotenv==0.10.1
pywin32-ctypes==0.2.0
requests==2.21.0
selenium==3.14.1
six==1.12.0
tqdm==4.29.0
typed-ast==1.3.1
urllib3==1.23
whichcraft==0.5.2
wrapt==1.10.11
